+++ example/app/ack-angular-webcam/AckMediaDevices.component.d.ts
@@ -1,0 +1,12 @@
+import { EventEmitter } from '@angular/core';
+export interface device {
+}
+export declare class AckMediaDevices {
+    array: device[];
+    arrayChange: EventEmitter<device[]>;
+    error: EventEmitter<Event>;
+    errorChange: EventEmitter<Event>;
+    catcher: EventEmitter<Event>;
+    ngOnInit(): void;
+    loadDevices(): Promise<device[]>;
+}