+++ example/app/ack-angular-webcam/index.d.ts
@@ -1,7 +1,8 @@
 import { WebCamComponent } from './webcam.component';
+import { WebCamModule } from './webcam.module';
 /**
  * @module
  * @description
  * Entry point from which you should import webcam component
  */
-export { WebCamComponent };
+export { WebCamComponent, WebCamModule };