+++ lib/index.ts
@@ -1,9 +1,5 @@
 import { WebCamComponent } from './webcam.component';
-// import { FallbackDispatcher } from './fallback/ts/fallback.dispatcher';
-
-// Following only works with webpack. Not AoT Compatible
-// import './lib/fallback/jscam_canvas_only.swf'
-// import './lib/fallback/jscam.swf'
+import { WebCamModule } from './webcam.module';
 
 /**
  * @module
@@ -11,6 +7,6 @@ import { WebCamComponent } from './webcam.component';
  * Entry point from which you should import webcam component
  */
 export {
-  WebCamComponent
-  // , FallbackDispatcher
+  WebCamComponent,
+  WebCamModule
 }